pyproject.toml
[tool.poetry]
name = "flac-converter"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
cloudconvert = "^2.0.0"

[tool.poetry.dev-dependencies]
pytest-mock-generator = "^1.0.0"
pytest = "^6.2.5"
pytest-mock = "^3.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
